+++ common/src/main/java/de/cech12/woodenhopper/block/WoodenHopperBlock.java
@@ -28,6 +28,7 @@ public class WoodenHopperBlock extends HopperBlock {
     }
 
     @Override
+    @NotNull
     public BlockEntity newBlockEntity(@NotNull BlockPos pos, @NotNull BlockState state) {
         return Services.REGISTRY.getNewBlockEntity(pos, state);
     }