+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [26.1-4.0.0.0] - 2026-03-28
+### Changed
+- updated to Minecraft 26.1 (Fabric 0.144.3+26.1, NeoForge 26.1.0.7-beta, Forge 62.0.4)
+- updated Cloth Config support (26.1.154) (Fabric/Quilt)
+- updated ModMenu support (18.0.0-alpha.8) (Fabric/Quilt)
+
 ## [1.21.11-3.8.0.0] - 2025-12-13
 ### Changed
 - updated to Minecraft 1.21.11 (Fabric 0.139.5+1.21.11, NeoForge 21.11.6-beta, Forge 61.0.2)